refactor(xtask): simplify perf harness surface
